rename: component와 view의 구분에 따른 파일 변경
