+++ src/view/purchaseAmountInput.js
