+++ src/view/purchaseLottoStatus.js
